fix: faster demo load and grid remount after tab switch
- Split grid demo into GridDemo.svelte so table mode avoids 10k-row init
- Defer full dataset via requestIdleCallback; show 250 rows immediately
- Reset virtualizer when scroll container unmounts; rebind refs on remount
- Share demo person generator; table demo uses lite 200-row dataset
- Husky pre-commit uses npm; bump vite to 7.3.x
- Add regression test for grid visibility after table roundtrip
