+++ .husky/pre-commit
@@ -1,3 +1,3 @@
 # Rebuild registry from source files and stage changes
-bun run registry:build
+npm run registry:build
 git add static/r/
