+++ package.json
@@ -40,7 +40,7 @@
 		"@sveltejs/adapter-auto": "^7.0.0",
 		"@sveltejs/adapter-vercel": "^6.2.0",
 		"@sveltejs/kit": "^2.48.5",
-		"@sveltejs/vite-plugin-svelte": "^6.2.1",
+		"@sveltejs/vite-plugin-svelte": "^6.2.4",
 		"@tailwindcss/vite": "^4.1.17",
 		"@vitest/browser-playwright": "^4.0.10",
 		"bits-ui": "^2.11.0",
@@ -57,7 +57,7 @@
 		"tailwindcss": "^4.1.17",
 		"tw-animate-css": "^1.4.0",
 		"typescript": "^5.9.3",
-		"vite": "^7.2.2",
+		"vite": "^7.3.3",
 		"vitest": "^4.0.10",
 		"vitest-browser-svelte": "^2.0.1"
 	},
