+++ src/routes/+page.svelte
@@ -1,435 +1,34 @@
 <script lang="ts">
-	import { faker } from '@faker-js/faker';
-	import type { ColumnDef } from '@tanstack/table-core';
+	import { Button, DataTableSkeleton, useWindowSize } from '$lib';
 	import {
-		Button,
-		DataGrid,
-		DataGridFilterMenu,
-		DataGridSortMenu,
-		DataGridRowHeightMenu,
-		DataGridViewMenu,
-		DataGridKeyboardShortcuts,
-		DataTableSkeleton,
-		getDataGridSelectColumn,
-		useDataGrid,
-		useDataGridUndoRedo,
-		useWindowSize,
-		getFilterFn,
-		type FileCellData,
-		type UndoRedoCellUpdate
-	} from '$lib';
+		departments,
+		generateDemoPeople,
+		skills,
+		statuses,
+		type DemoPerson
+	} from '$lib/demo/person-data.js';
 	import { Toaster } from '$lib/components/ui/sonner';
 	import ModeToggle from '$lib/components/mode-toggle.svelte';
 	import DataTableShowcase from './data-table-showcase.svelte';
+	import GridDemo from './grid-demo.svelte';
 
 	type DemoMode = 'grid' | 'table';
 	type TableMode = 'basic' | 'advanced';
 	type AdvancedFilterUi = 'advancedFilters' | 'commandFilters';
 
-	interface Person {
-		id: string;
-		name?: string;
-		age?: number;
-		email?: string;
-		website?: string;
-		notes?: string;
-		salary?: number;
-		department?: string;
-		status?: string;
-		skills?: string[];
-		isActive?: boolean;
-		startDate?: string;
-		attachments?: FileCellData[];
-	}
-
-	faker.seed(12345);
-
-	const departments = ['Engineering', 'Marketing', 'Sales', 'HR', 'Finance'] as const;
-	const statuses = ['Active', 'On Leave', 'Remote', 'In Office'] as const;
-	const skills = [
-		'JavaScript',
-		'TypeScript',
-		'React',
-		'Node.js',
-		'Python',
-		'SQL',
-		'AWS',
-		'Docker',
-		'Git',
-		'Agile'
-	] as const;
-
-	const notes = [
-		'Excellent team player with strong communication skills. Consistently meets deadlines and delivers high-quality work.',
-		'Currently working on the Q4 project initiative. Requires additional training in advanced analytics tools.',
-		'Relocated from the Seattle office last month. Adjusting well to the new team dynamics and company culture.',
-		'Submitted request for professional development courses. Shows great initiative in learning new technologies.',
-		'Outstanding performance in the last quarter. Recommended for leadership training program next year.',
-		'Recently completed certification in project management. Looking to take on more responsibility in upcoming projects.',
-		'Needs improvement in time management. Working with mentor to develop better organizational skills.',
-		'Transferred from the marketing department. Bringing valuable cross-functional experience to the team.',
-		'On track for promotion consideration. Has exceeded expectations in client relationship management.',
-		'Participating in the company mentorship program. Showing strong potential for career advancement.',
-		'Recently returned from parental leave. Successfully reintegrated into current project workflows.',
-		'Fluent in three languages. Often assists with international client communications and translations.',
-		'Leading the diversity and inclusion initiative. Organizing monthly team building events and workshops.',
-		'Requested flexible work arrangement for family care. Maintaining productivity while working remotely.',
-		"Completed advanced training in data visualization. Now serving as the team's go-to expert for dashboards."
-	];
-
-	const sampleFiles = [
-		{ name: 'Resume.pdf', type: 'application/pdf', sizeRange: [50, 500] },
-		{ name: 'Contract.pdf', type: 'application/pdf', sizeRange: [100, 300] },
-		{ name: 'ID_Document.pdf', type: 'application/pdf', sizeRange: [200, 400] },
-		{ name: 'Profile_Photo.jpg', type: 'image/jpeg', sizeRange: [500, 2000] },
-		{
-			name: 'Presentation.pptx',
-			type: 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
-			sizeRange: [1000, 5000]
-		},
-		{
-			name: 'Report.docx',
-			type: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-			sizeRange: [100, 800]
-		},
-		{
-			name: 'Timesheet.xlsx',
-			type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-			sizeRange: [50, 200]
-		},
-		{ name: 'Certificate.pdf', type: 'application/pdf', sizeRange: [200, 500] },
-		{ name: 'Background_Check.pdf', type: 'application/pdf', sizeRange: [300, 600] },
-		{ name: 'Training_Video.mp4', type: 'video/mp4', sizeRange: [5000, 15000] }
-	] as const;
-
-	function generatePerson(id: number): Person {
-		const firstName = faker.person.firstName();
-		const lastName = faker.person.lastName();
-
-		// Generate 0-3 files for this person
-		const fileCount = faker.number.int({ min: 0, max: 3 });
-		const selectedFiles = faker.helpers.arrayElements(sampleFiles, fileCount);
-
-		const attachments: FileCellData[] = selectedFiles.map((file, index) => {
-			const sizeKB = faker.number.int({
-				min: file.sizeRange[0],
-				max: file.sizeRange[1]
-			});
-			return {
-				id: `${id}-file-${index}`,
-				name: file.name,
-				size: sizeKB * 1024, // Convert to bytes
-				type: file.type,
-				url: `https://example.com/files/${id}/${file.name}`
-			};
-		});
-
-		return {
-			id: faker.string.nanoid(8),
-			name: `${firstName} ${lastName}`,
-			age: faker.number.int({ min: 22, max: 65 }),
-			email: faker.internet.email({ firstName, lastName }).toLowerCase(),
-			website: faker.internet.url().replace(/\/$/, ''),
-			notes: faker.helpers.arrayElement(notes),
-			salary: faker.number.int({ min: 40000, max: 150000 }),
-			department: faker.helpers.arrayElement(departments),
-			status: faker.helpers.arrayElement(statuses),
-			isActive: faker.datatype.boolean(),
-			startDate:
-				faker.date.between({ from: '2018-01-01', to: '2024-01-01' }).toISOString().split('T')[0] ??
-				'',
-			skills: faker.helpers.arrayElements(skills, { min: 1, max: 5 }),
-			attachments
-		};
-	}
-
-	const initialData: Person[] = Array.from({ length: 10000 }, (_, i) => generatePerson(i + 1));
-
-	let data = $state<Person[]>(initialData);
 	let demoMode = $state<DemoMode>('grid');
 	let tableMode = $state<TableMode>('basic');
 	let advancedFilterUi = $state<AdvancedFilterUi>('commandFilters');
 	let showTableSkeleton = $state(false);
 
-	const { trackCellsUpdate, trackRowsAdd, trackRowsDelete } = useDataGridUndoRedo({
-		data: () => data,
-		onDataChange: (newData) => {
-			data = newData;
-		},
-		getRowId: (row) => row.id
-	});
+	// Table demo only needs a small slice — avoid generating 10k rows on first paint.
+	const tableRows = $state<DemoPerson[]>(generateDemoPeople(200, { lite: true }));
 
-	// Dynamic height based on window size
 	const windowSize = useWindowSize({ defaultHeight: 760 });
 	const gridHeight = $derived(Math.max(400, windowSize.height - 150));
-
-	const filterFn = getFilterFn<Person>();
-	const tableRows = $derived(data.slice(0, 200));
-
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	const columns: ColumnDef<Person, any>[] = [
-		getDataGridSelectColumn<Person>({ enableRowMarkers: true }),
-		{
-			id: 'name',
-			accessorKey: 'name',
-			header: 'Name',
-			minSize: 180,
-			filterFn,
-			meta: {
-				label: 'Name',
-				cell: {
-					variant: 'short-text'
-				}
-			}
-		},
-		{
-			id: 'age',
-			accessorKey: 'age',
-			header: 'Age',
-			minSize: 100,
-			filterFn,
-			meta: {
-				label: 'Age',
-				cell: {
-					variant: 'number',
-					min: 18,
-					max: 100,
-					step: 1
-				}
-			}
-		},
-		{
-			id: 'email',
-			accessorKey: 'email',
-			header: 'Email',
-			minSize: 240,
-			filterFn,
-			meta: {
-				label: 'Email',
-				cell: {
-					variant: 'short-text'
-				}
-			}
-		},
-		{
-			id: 'website',
-			accessorKey: 'website',
-			header: 'Website',
-			minSize: 240,
-			filterFn,
-			meta: {
-				label: 'Website',
-				cell: {
-					variant: 'url'
-				}
-			}
-		},
-		{
-			id: 'notes',
-			accessorKey: 'notes',
-			header: 'Notes',
-			minSize: 200,
-			filterFn,
-			meta: {
-				label: 'Notes',
-				cell: {
-					variant: 'long-text'
-				}
-			}
-		},
-		{
-			id: 'salary',
-			accessorKey: 'salary',
-			header: 'Salary',
-			minSize: 180,
-			filterFn,
-			meta: {
-				label: 'Salary',
-				cell: {
-					variant: 'number',
-					min: 0,
-					step: 1000
-				}
-			}
-		},
-		{
-			id: 'department',
-			accessorKey: 'department',
-			header: 'Department',
-			minSize: 180,
-			filterFn,
-			meta: {
-				label: 'Department',
-				cell: {
-					variant: 'select',
-					options: departments.map((dept) => ({
-						label: dept,
-						value: dept
-					}))
-				}
-			}
-		},
-		{
-			id: 'status',
-			accessorKey: 'status',
-			header: 'Status',
-			minSize: 180,
-			filterFn,
-			meta: {
-				label: 'Status',
-				cell: {
-					variant: 'select',
-					options: statuses.map((status) => ({
-						label: status,
-						value: status
-					}))
-				}
-			}
-		},
-		{
-			id: 'skills',
-			accessorKey: 'skills',
-			header: 'Skills',
-			minSize: 240,
-			filterFn,
-			meta: {
-				label: 'Skills',
-				cell: {
-					variant: 'multi-select',
-					options: skills.map((skill) => ({
-						label: skill,
-						value: skill
-					}))
-				}
-			}
-		},
-		{
-			id: 'isActive',
-			accessorKey: 'isActive',
-			header: 'Active',
-			minSize: 140,
-			filterFn,
-			meta: {
-				label: 'Active',
-				cell: {
-					variant: 'checkbox'
-				}
-			}
-		},
-		{
-			id: 'startDate',
-			accessorKey: 'startDate',
-			header: 'Start Date',
-			minSize: 150,
-			filterFn,
-			meta: {
-				label: 'Start Date',
-				cell: {
-					variant: 'date'
-				}
-			}
-		},
-		{
-			id: 'attachments',
-			accessorKey: 'attachments',
-			header: 'Attachments',
-			minSize: 240,
-			filterFn,
-			meta: {
-				label: 'Attachments',
-				cell: {
-					variant: 'file',
-					maxFileSize: 10 * 1024 * 1024, // 10MB
-					maxFiles: 5,
-					accept: 'image/*,video/*,audio/*,.pdf,.doc,.docx,.xls,.xlsx,.ppt,.pptx',
-					multiple: true
-				}
-			}
-		}
-	];
-
-	function onRowAdd() {
-		const newRow = {
-			id: faker.string.nanoid(8)
-		};
-
-		data = [...data, newRow];
-		trackRowsAdd([newRow]);
-
-		return {
-			rowIndex: data.length - 1,
-			columnId: 'name'
-		};
-	}
-
-	function onRowsAdd(count: number) {
-		const newRows = Array.from({ length: count }, () => ({
-			id: faker.string.nanoid(8)
-		}));
-		data = [...data, ...newRows];
-		trackRowsAdd(newRows);
-	}
-
-	function onRowsDelete(rows: Person[]) {
-		trackRowsDelete(rows);
-		data = data.filter((row) => !rows.includes(row));
-	}
-
-
-	function onDataChange(newData: Person[]) {
-		data = newData;
-	}
-
-	async function onFilesUpload({ files }: { files: File[] }) {
-		// Simulate upload delay
-		await new Promise((resolve) => setTimeout(resolve, 800));
-
-		return files.map((file) => ({
-			id: crypto.randomUUID(),
-			name: file.name,
-			size: file.size,
-			type: file.type,
-			url: URL.createObjectURL(file)
-		}));
-	}
-
-	async function onFilesDelete({
-		fileIds,
-		rowIndex,
-		columnId
-	}: {
-		fileIds: string[];
-		rowIndex: number;
-		columnId: string;
-	}) {
-		console.log(
-			`Deleting ${fileIds.length} file(s) from row ${rowIndex}, column ${columnId}:`,
-			fileIds
-		);
-	}
-
-	const { table, ...dataGridProps } = useDataGrid({
-		columns,
-		data: () => data, // Pass as getter for Svelte 5 reactivity
-		onDataChange,
-		onRowAdd,
-		onRowsAdd,
-		onRowsDelete,
-		onFilesUpload,
-		onFilesDelete,
-		getRowId: (row) => row.id,
-		initialState: {
-			columnPinning: {
-				left: ['select']
-			}
-		},
-		enableSearch: true,
-		enablePaste: true
-	});
 </script>
 
 <div class="mx-auto flex w-full max-w-[1600px] flex-col gap-4 px-6 py-6">
-	<!-- Header -->
 	<div class="flex flex-col gap-2">
 		<div class="flex items-center justify-between">
 			<div>
@@ -474,23 +73,7 @@
 	</div>
 
 	{#if demoMode === 'grid'}
-		<!-- Toolbar -->
-		<div role="toolbar" aria-orientation="horizontal" class="flex items-center justify-between">
-			<DataGridKeyboardShortcuts
-				enableSearch={!!dataGridProps.searchState}
-				enableUndoRedo
-				enablePaste
-			/>
-			<div class="flex items-center gap-2">
-				<DataGridFilterMenu {table} />
-				<DataGridSortMenu {table} />
-				<DataGridRowHeightMenu {table} />
-				<DataGridViewMenu {table} />
-			</div>
-		</div>
-
-		<!-- Data Grid -->
-		<DataGrid {...dataGridProps} {table} height={gridHeight} />
+		<GridDemo height={gridHeight} />
 	{:else}
 		<div
 			class="flex flex-wrap items-center justify-between gap-3 rounded-lg border bg-muted/30 p-3"
